jeopardy-service/requirements.txt
bs4
flask
flask-cors
pandas
pyodbc
azure-identity
azure-storage-blob
requests